Backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
Pillowdjangorestframework-simplejwt==5.3.1
